extract and consolidate existing type guards
