+++ frontend/src/metabase-types/guards/data-apps.ts
@@ -1,7 +1,17 @@
-import type { ActionDashboardCard } from "metabase-types/api";
+import type {
+  ActionDashboardCard,
+  Dashboard,
+  DataAppNavItem,
+} from "metabase-types/api";
 
 export function isActionDashboardCard(
   dashcard: unknown,
 ): dashcard is ActionDashboardCard {
   return "action" in (dashcard as ActionDashboardCard);
+}
+
+export function isNavItem(
+  object: Dashboard | DataAppNavItem,
+): object is DataAppNavItem {
+  return "page_id" in object;
 }
