+++ frontend/src/metabase-types/guards/dom.ts
@@ -1,0 +1,3 @@
+export function isElement(a: any): a is Element {
+  return a instanceof Element;
+}
