+++ frontend/src/metabase-types/guards/forms-legacy.ts
@@ -1,0 +1,14 @@
+import type {
+  StandardFormFieldDefinition,
+  CustomFormFieldDefinition,
+  FormFieldDefinition,
+} from "metabase-types/forms/legacy";
+
+export function isCustomWidget(
+  formField: FormFieldDefinition,
+): formField is CustomFormFieldDefinition {
+  return (
+    !(formField as StandardFormFieldDefinition).type &&
+    typeof (formField as CustomFormFieldDefinition).widget === "function"
+  );
+}
