+++ frontend/src/metabase/components/form/FormikCustomForm/CustomFormField.tsx
@@ -2,6 +2,7 @@ import React, { useCallback, useMemo } from "react";
 import { getIn } from "icepick";
 import _ from "underscore";
 
+import { isCustomWidget } from "metabase-types/guards/forms";
 import FormField from "metabase/components/form/FormikFormField";
 import FormWidget from "metabase/components/form/FormWidget";
 
@@ -11,20 +12,10 @@ import { useOnUnmount } from "metabase/hooks/use-on-unmount";
 import {
   BaseFieldDefinition,
   StandardFormFieldDefinition,
-  CustomFormFieldDefinition,
   FormFieldDefinition,
 } from "metabase-types/forms";
 
 import { useForm } from "./context";
-
-function isCustomWidget(
-  formField: FormFieldDefinition,
-): formField is CustomFormFieldDefinition {
-  return (
-    !(formField as StandardFormFieldDefinition).type &&
-    typeof (formField as CustomFormFieldDefinition).widget === "function"
-  );
-}
 
 export interface CustomFormFieldProps extends BaseFieldDefinition {
   onChange?: (e: unknown) => void;
