+++ frontend/src/metabase/parameters/utils/formatting.ts
@@ -1,10 +1,8 @@
 import { ngettext, msgid } from "ttag";
 
 import { formatValue } from "metabase/lib/formatting";
-import {
-  getParameterType,
-  isFieldFilterParameter,
-} from "metabase-lib/parameters/utils/parameter-type";
+import { isFieldFilterParameter } from "metabase-types/guards/parameters";
+import { getParameterType } from "metabase-lib/parameters/utils/parameter-type";
 
 import { UiParameter } from "metabase-lib/parameters/types";
 import { formatDateValue } from "./date-formatting";
