+++ frontend/src/metabase/parameters/utils/parameter-type.ts
@@ -1,21 +1,10 @@
 import _ from "underscore";
 import { Parameter } from "metabase-types/types/Parameter";
-import { FieldFilterUiParameter } from "metabase-lib/parameters/types";
-import {
-  FIELD_FILTER_PARAMETER_TYPES,
-  SINGLE_OR_MULTI_SELECTABLE_TYPES,
-} from "metabase-lib/parameters/constants";
+import { SINGLE_OR_MULTI_SELECTABLE_TYPES } from "metabase-lib/parameters/constants";
 import {
   getParameterType,
   getParameterSubType,
 } from "metabase-lib/parameters/utils/parameter-type";
-
-export function isFieldFilterParameter(
-  parameter: Parameter,
-): parameter is FieldFilterUiParameter {
-  const type = getParameterType(parameter);
-  return FIELD_FILTER_PARAMETER_TYPES.includes(type);
-}
 
 export function isSingleOrMultiSelectable(parameter: Parameter): boolean {
   const type: string = getParameterType(parameter);
