+++ frontend/src/metabase/query_builder/actions/core/initializeQB.ts
@@ -18,7 +18,8 @@ import {
   GetState,
   QueryBuilderUIControls,
 } from "metabase-types/store";
-import { Card, SavedCard } from "metabase-types/types/Card";
+import type { Card } from "metabase-types/types/Card";
+import { isSavedCard } from "metabase-types/guards/card";
 import { isNotNull } from "metabase/core/utils/array";
 import { cardIsEquivalent } from "metabase-lib/queries/utils/card";
 import { normalize } from "metabase-lib/queries/utils/normalize";
@@ -195,10 +196,6 @@ function parseHash(hash?: string) {
   }
 
   return { options, serializedCard };
-}
-
-function isSavedCard(card: Card): card is SavedCard {
-  return !!(card as SavedCard).id;
 }
 
 export const INITIALIZE_QB = "metabase/qb/INITIALIZE_QB";
